chore: remove use client and css from files not need, move src/app/_components and _components of about-page to src/components, remove h2 in css
