+++ src/app/about/page.tsx
@@ -1,25 +1,13 @@
-import "@/styles/globals.css";
-import { Header } from "../_components/Header";
-import { IBM_Plex_Sans_Thai, Inter } from "next/font/google";
-import { AboutSession } from "./_components/AboutSession";
-import { ProjectsSession } from "./_components/ProjectsSession";
-import { MemberSession } from "./_components/MemberSession";
-import { BackButton } from "../_components/BackButton";
 
-const inter = Inter({
-  subsets: ["latin"],
-  variable: "--font-inter",
-});
-
-const ibmPlexSansThai = IBM_Plex_Sans_Thai({
-  subsets: ["latin"],
-  variable: "--font-ibm-plex-sans-thai",
-  weight: ["400", "500", "600", "700"],
-});
+import { Header } from "../../components/common/Header";
+import { AboutSession } from "../../components/about/AboutSession";
+import { ProjectsSession } from "../../components/about/ProjectsSession";
+import { MemberSession } from "../../components/about/MemberSession";
+import { BackButton } from "../../components/common/Back-Button";
 
 export default function RootHead() {
   return (
-    <html lang="en" className={`${inter.variable} ${ibmPlexSansThai.variable}`}>
+    <>
       <body>
         <Header />
         <BackButton backUrl="/"/>
@@ -27,6 +15,6 @@ export default function RootHead() {
         <ProjectsSession />
         <MemberSession />
       </body>
-    </html>
+    </>
   );
 }
