+++ src/styles/globals.css
@@ -71,10 +71,3 @@
     @apply bg-background text-foreground;
   }
 }
-
-h2 {
-  font-weight: 600;
-  font-size: 30px;
-  line-height: 36px;
-  letter-spacing: -0.0075em;
-}
